Add civic-service Phase 2: prosperity/luxury buildings close the satisfaction loop
Continues Phase 1's tiered-service foundation (Electricity/Water) with the prosperity
(Hospital/Fire/Sanitation/Cemetery) and luxury (Park/Museum) categories that were designed
conceptually before the tier pivot but deferred. Unlike Electricity/Water, these have no natural
"capacity" unit (spec §20) — instead they close a loop open since original Fase 2:
BuildingRegistry.satisfactionPercent was computed every tick but never consumed by anything.

9 new BuildingType constants in 5 categories: Healthcare (Clinic->Hospital), Fire (Volunteer Fire
Brigade->Fire Station), Sanitation (Waste Collection->Incinerator, modeled as one category's two
tiers rather than two separate ones), and Cemetery/Park/Museum with no tier 2 — the user described
these as services that "already grant the maximum level" once built. Museum is the one civic
building with its own revenue (tourism), confirmed during Phase 1's scoping questions.

UtilitySystem gains 6 more flood-fill categories (coverage bit only, no capacity/demand ratio —
there's no sensible "population served" number for a hospital). PopulationSystem replaces the flat
+/-satisfaction step with movement toward a ceiling derived from which prosperity/luxury bits cover
a tile (60 base / 85 prosperity / 100 luxury), and growResidential/growWorkplace now multiply
growth by satisfactionPercent/100 on top of Difficulty and the Phase 1 capacity ratios — a city
with zero prosperity coverage stays capped near 60% growth multiplier permanently.

New BuildCivicBuildingCommand (data-driven, like BuildProductionBuildingCommand) reuses Phase 1's
REJECTED_SERVICE_TIER_LOCKED/REJECTED_INSUFFICIENT_FUNDS — no new CommandResult needed. Extracted
BuildingRegistry.residentialPopulationOfCity as a public method shared with
AbstractPlaceUtilityBuildingCommand instead of a private duplicate.

9 new service bits push serviceFlags past a byte's 8 bits, as anticipated in Phase 1: Chunk.
serviceFlags widens from byte[] to int[], bumping ChunkDeltaIO.FORMAT_VERSION to 3 (old saves
unreadable, same precedent as prior format bumps this session, no release yet).

Adds unit/command/system test coverage and documents the design, including what's still explicitly
out of scope (city hall, central bank, police/education/churches, TheOtown-style building density
evolution) in docs/roadmap.md and CLAUDE.md.

+++ simulation-core/src/main/java/com/kosmos/atlas/sim/world/Chunk.java
@@ -27,7 +27,8 @@ public final class Chunk {
     public final byte[] zoneType = new byte[WorldConstants.TILES_PER_CHUNK];
     public final byte[] roadType = new byte[WorldConstants.TILES_PER_CHUNK];
     public final int[] buildingId = new int[WorldConstants.TILES_PER_CHUNK];
-    public final byte[] serviceFlags = new byte[WorldConstants.TILES_PER_CHUNK];
+    /** {@code int}, not {@code byte} — Fase 2's civic-service tiers pushed past 8 usable bits. */
+    public final int[] serviceFlags = new int[WorldConstants.TILES_PER_CHUNK];
 
     private int chunkX;
     private int chunkY;
@@ -55,7 +56,7 @@ public final class Chunk {
         Arrays.fill(zoneType, WorldConstants.ZONE_NONE);
         Arrays.fill(roadType, WorldConstants.ROAD_NONE);
         Arrays.fill(buildingId, WorldConstants.NO_BUILDING);
-        Arrays.fill(serviceFlags, (byte) 0);
+        Arrays.fill(serviceFlags, 0);
     }
 
     public int chunkX() {
